Refactor: Move free helper functions out of use-escrow-transaction.ts (closes #960) (#965)
* refactor: extract escrow helper functions from hook (closes #960)

* refactor: address code review feedback — arrow functions, explicit return types, remove duplicate toast
